fix: remove suporte a refs externas no sandcastle
